+++ tests/tests/constants/testContracts.ts
@@ -294,3 +294,101 @@ export const ERC20_BYTECODE =
   "65207a65726f206164647265737345524332303a2064656372656173656420616c6c6f77616e63652062656c6f7720" +
   "7a65726fa265627a7a723158208bfd5c482ccbde97b679a417e6f0a1698bb5491e25bb256602f6c4271b11757b6473" +
   "6f6c63430005110032";
+
+export const TEST_SUBSCRIPTION_CONTRACT_BYTECODE =
+  "0x608060405234801561001057600080fd5b50610041337fffffffffffffffffffffffffffffffffffffffffff" +
+  "ffffffffffffffffffffff61004660201b60201c565b610291565b600073ffffffffffffffffffffffffffffff" +
+  "ffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614156100e9576040517f08c379a00000" +
+  "0000000000000000000000000000000000000000000000000000815260040180806020018281038252601f8152" +
+  "602001807f45524332303a206d696e7420746f20746865207a65726f2061646472657373008152506020019150" +
+  "5060405180910390fd5b6101028160025461020960201b610c7c1790919060201c565b60028190555061015d81" +
+  "6000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffff" +
+  "ffff1681526020019081526020016000205461020960201b610c7c1790919060201c565b6000808473ffffffff" +
+  "ffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190" +
+  "8152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffff" +
+  "ffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523" +
+  "b3ef836040518082815260200191505060405180910390a35050565b6000808284019050838110156102875760" +
+  "40517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001" +
+  "8281038252601b8152602001807f536166654d6174683a206164646974696f6e206f766572666c6f7700000000" +
+  "0081525060200191505060405180910390fd5b8091505092915050565b610e3a806102a06000396000f3fe6080" +
+  "60405234801561001057600080fd5b50600436106100885760003560e01c806370a082311161005b57806370a0" +
+  "8231146101fd578063a457c2d714610255578063a9059cbb146102bb578063dd62ed3e1461032157610088565b" +
+  "8063095ea7b31461008d57806318160ddd146100f357806323b872dd146101115780633950935114610197575b" +
+  "600080fd5b6100d9600480360360408110156100a357600080fd5b81019080803573ffffffffffffffffffffff" +
+  "ffffffffffffffffff16906020019092919080359060200190929190505050610399565b604051808215151515" +
+  "815260200191505060405180910390f35b6100fb6103b7565b6040518082815260200191505060405180910390" +
+  "f35b61017d6004803603606081101561012757600080fd5b81019080803573ffffffffffffffffffffffffffff" +
+  "ffffffffffff169060200190929190803573ffffffffffffffffffffffffffffffffffffffff16906020019092" +
+  "9190803590602001909291905050506103c1565b604051808215151515815260200191505060405180910390f3" +
+  "5b6101e3600480360360408110156101ad57600080fd5b81019080803573ffffffffffffffffffffffffffffff" +
+  "ffffffffff1690602001909291908035906020019092919050505061049a565b60405180821515151581526020" +
+  "0191505060405180910390f35b61023f6004803603602081101561021357600080fd5b81019080803573ffffff" +
+  "ffffffffffffffffffffffffffffffffff16906020019092919050505061054d565b6040518082815260200191" +
+  "505060405180910390f35b6102a16004803603604081101561026b57600080fd5b81019080803573ffffffffff" +
+  "ffffffffffffffffffffffffffffff16906020019092919080359060200190929190505050610595565b604051" +
+  "808215151515815260200191505060405180910390f35b610307600480360360408110156102d157600080fd5b" +
+  "81019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190803590602001909291" +
+  "90505050610662565b604051808215151515815260200191505060405180910390f35b61038360048036036040" +
+  "81101561033757600080fd5b81019080803573ffffffffffffffffffffffffffffffffffffffff169060200190" +
+  "929190803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050610680565b6040" +
+  "518082815260200191505060405180910390f35b60006103ad6103a6610707565b848461070f565b6001905092" +
+  "915050565b6000600254905090565b60006103ce848484610906565b61048f846103da610707565b61048a8560" +
+  "4051806060016040528060288152602001610d7060289139600160008b73ffffffffffffffffffffffffffffff" +
+  "ffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600061" +
+  "0440610707565b73ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffff" +
+  "ffffffffff16815260200190815260200160002054610bbc9092919063ffffffff16565b61070f565b60019050" +
+  "9392505050565b60006105436104a7610707565b8461053e85600160006104b8610707565b73ffffffffffffff" +
+  "ffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260" +
+  "200160002060008973ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffff" +
+  "ffffffffffff16815260200190815260200160002054610c7c90919063ffffffff16565b61070f565b60019050" +
+  "92915050565b60008060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffff" +
+  "ffffffffffffffffffff168152602001908152602001600020549050919050565b60006106586105a261070756" +
+  "5b8461065385604051806060016040528060258152602001610de160259139600160006105cc610707565b73ff" +
+  "ffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260" +
+  "200190815260200160002060008a73ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffff" +
+  "ffffffffffffffffffffffff16815260200190815260200160002054610bbc9092919063ffffffff16565b6107" +
+  "0f565b6001905092915050565b600061067661066f610707565b8484610906565b6001905092915050565b6000" +
+  "600160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffff" +
+  "ffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ff" +
+  "ffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b6000" +
+  "33905090565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffff" +
+  "ffffffffffffff161415610795576040517f08c379a00000000000000000000000000000000000000000000000" +
+  "00000000008152600401808060200182810382526024815260200180610dbd6024913960400191505060405180" +
+  "910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffff" +
+  "ffffffffffff16141561081b576040517f08c379a0000000000000000000000000000000000000000000000000" +
+  "000000008152600401808060200182810382526022815260200180610d28602291396040019150506040518091" +
+  "0390fd5b80600160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffff" +
+  "ffffffffffffffff16815260200190815260200160002060008473ffffffffffffffffffffffffffffffffffff" +
+  "ffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508173" +
+  "ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f" +
+  "8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b92583604051808281526020019150" +
+  "5060405180910390a3505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffff" +
+  "ffffffffffffffffffffffffffffff16141561098c576040517f08c379a0000000000000000000000000000000" +
+  "000000000000000000000000008152600401808060200182810382526025815260200180610d98602591396040" +
+  "0191505060405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffff" +
+  "ffffffffffffffffffffffffffff161415610a12576040517f08c379a000000000000000000000000000000000" +
+  "0000000000000000000000008152600401808060200182810382526023815260200180610d0560239139604001" +
+  "91505060405180910390fd5b610a7d81604051806060016040528060268152602001610d4a6026913960008087" +
+  "73ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681" +
+  "5260200190815260200160002054610bbc9092919063ffffffff16565b6000808573ffffffffffffffffffffff" +
+  "ffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000" +
+  "2081905550610b10816000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffff" +
+  "ffffffffffffffffffffff16815260200190815260200160002054610c7c90919063ffffffff16565b60008084" +
+  "73ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681" +
+  "52602001908152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff168373ffffff" +
+  "ffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f5" +
+  "5a4df523b3ef836040518082815260200191505060405180910390a3505050565b6000838311158290610c6957" +
+  "6040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020" +
+  "01828103825283818151815260200191508051906020019080838360005b83811015610c2e5780820151818401" +
+  "52602081019050610c13565b50505050905090810190601f168015610c5b578082038051600183602003610100" +
+  "0a031916815260200191505b509250505060405180910390fd5b5060008385039050809150509392505050565b" +
+  "600080828401905083811015610cfa576040517f08c379a0000000000000000000000000000000000000000000" +
+  "00000000000000815260040180806020018281038252601b8152602001807f536166654d6174683a2061646469" +
+  "74696f6e206f766572666c6f77000000000081525060200191505060405180910390fd5b809150509291505056" +
+  "fe45524332303a207472616e7366657220746f20746865207a65726f206164647265737345524332303a206170" +
+  "70726f766520746f20746865207a65726f206164647265737345524332303a207472616e7366657220616d6f75" +
+  "6e7420657863656564732062616c616e636545524332303a207472616e7366657220616d6f756e742065786365" +
+  "65647320616c6c6f77616e636545524332303a207472616e736665722066726f6d20746865207a65726f206164" +
+  "647265737345524332303a20617070726f76652066726f6d20746865207a65726f206164647265737345524332" +
+  "303a2064656372656173656420616c6c6f77616e63652062656c6f77207a65726fa265627a7a72315820c7a5ff" +
+  "abf642bda14700b2de42f8c57b36621af020441df825de45fd2b3e1c5c64736f6c63430005100032";
